% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iCARE}
\alias{iCARE}
\title{Example dataset dataset for iCARE study}
\format{
A longitudinal dataset in wide format, i.e one row per person, one column per variable.
}
\usage{
data(iCARE)
}
\description{
Example dataset dataset for iCARE study to illustrate how the functions in the package work
}
\examples{
# Load data into global environment
data(iCARE)
}
\keyword{dataset}
