% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{inspect_missing}
\alias{inspect_missing}
\title{Inspect missingness}
\usage{
inspect_missing(data, min_value = 0, filter_var = NULL, filter_value = NULL)
}
\arguments{
\item{data}{Data object}

\item{min_value}{do you to see all, or say just those vars that have missigness above 0.2 - can specify, default = 0}

\item{filter_var}{specify variable for filtering}

\item{filter_value}{specify value for filtering in filter_var}
}
\value{

}
\description{
Inspect missingness
}
\examples{
inspect_missing(iCARE, min_value = .9)
}
