% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pie_chart}
\alias{plot_pie_chart}
\title{Plot pie chart}
\usage{
plot_pie_chart(
  data,
  variable,
  filter_response = NULL,
  title = NULL,
  filllab = NULL,
  labels = NULL,
  number_font_size = 4,
  return = c("plot", "data")
)
}
\arguments{
\item{data}{specify data file i.e. global_raw}

\item{variable}{specify variable you want in pie chart}

\item{filter_response}{filter response options e.g. c(1:4) to get rid of other values - cleaner plot}

\item{title}{specify title}

\item{filllab}{title = "An awesome pie chart"}

\item{labels}{use to relabel value e.g. c("Male", "Female")}

\item{number_font_size}{respecify the font size in case it doesnt fit}

\item{return}{do you want to get the data or plot}
}
\value{

}
\description{
Plot pie chart
}
\examples{
plot_pie_chart(data = iCARE,
               variable = sex,
               filter_response=c(1:2),
               title = "Sex distribution",
               labels = c("Male", "Female"),
               number_font_size = 6,
               return = "plot")
}
