% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_time_trend}
\alias{plot_time_trend}
\title{Plot time trend}
\usage{
plot_time_trend(
  data,
  variables,
  time_variable,
  filter_value = 1,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  var_names = NULL,
  max_value = 100,
  color = 0.5,
  return = c("data", "plot")
)
}
\arguments{
\item{data}{Data object}

\item{variables}{which variables do you want to see}

\item{time_variable}{choose between month or wave}

\item{filter_value}{which value of response e.g. 1 - most of the time}

\item{title}{specify title}

\item{xlab}{specify xlab}

\item{ylab}{specify ylab}

\item{var_names}{to relabel the header on the plot}

\item{max_value}{adjust the y axis}

\item{color}{change the color, available range = 0-1}

\item{return}{do you want to get data or plot -- I like this to}
}
\value{

}
\description{
Plot time trend
}
\examples{
plot_time_trend(data = iCARE,
                variables = c(impacvd_sq003, impacvd_sq001, impacvd_sq002),
                filter_value = 1,
                time_variable = month, # test out for month if you wish
                var_names = c("impacvd_sq003" = "Depression",
                              "impacvd_sq001" = "Anxiety",
                              "impacvd_sq002" = "Loneliness"),
                max_value = 25,
                title = "Trend over waves of survey",
                return = "plot")
}
